Rename _distinct_list.py to _link_set.py
